+++ packages/component/src/Composer.js
@@ -273,6 +273,7 @@ class Composer extends React.Component {
       //       If we let it thru, the code below become simplified and the user can plug in whatever they want for context, via Composer.props
       {
         activityRenderer,
+        // TODO: [P3] We should move adaptiveCardHostConfig to bundle
         adaptiveCardHostConfig: adaptiveCardHostConfig || defaultAdaptiveCardHostConfig(this.props.styleOptions),
         attachmentRenderer,